requirements.txt
dnspython
requests
termcolor